Create CrossVCMigration
This script was used for a cross virtual centre migration and consolidation from 2 VCs down to 1.

We were using the same shared storage on the source as the destination which is what we were pulling the datastore name from the old host and applying it on the destination. This removed the need for a storage vmotion and sped up migration.

Also included as part of this was defining backups on the VMs using tags and applying the appropriate tag on the destination VM. This could have been done more dynamically based on naming conventions but at the time I needed to just get the process working.

To be clear I am a pretty basic script monkey so I am sure all of this can be done better but feel free to use and improve.